Work in progress

+++ common/jobs/scm/basic-compose-pod-pipeline.groovy
@@ -20,10 +20,6 @@ node {
       CHECKOUT_COMMIT_ID="$(curl -Ssf http://127.0.0.1:8500/v1/kv/${PLATFORM_ENVIRONMENT}/${POD_NAME}/checkout_commit_id?raw)"
       PREVIOUS_POD_TAG="$(curl -Ss ${CONSUL_HTTP_ADDR}/v1/kv/platform-data/${PLATFORM_ENVIRONMENT}/${POD_NAME}/tag_version?raw)"
       POD_TAG="${CHECKOUT_COMMIT_ID:0:7}"
-      if [[ "${POD_TAG}" == "${PREVIOUS_POD_TAG}" ]]; then
-        echo [warning] commit id is the same, will not build again!
-        exit 0
-      fi
       VAULT_ADDR="$(curl -Ssf ${CONSUL_HTTP_ADDR}/v1/kv/platform-settings/vault_address?raw)"
       REGISTRY_ADDRESS="$(curl -Ssf ${CONSUL_HTTP_ADDR}/v1/kv/platform-settings/docker_registry_address?raw)"
       REGISTRY_PATH="$(curl -Ssf ${CONSUL_HTTP_ADDR}/v1/kv/platform-settings/docker_registry_path?raw)"
@@ -44,6 +40,10 @@ node {
       ansible all -i localhost, --connection=local -m template -a "src=nomad-job.hcl.j2 dest=nomad-job.hcl" >/dev/null
       nomad validate nomad-job.hcl
       nomad run -output nomad-job.hcl > nomad-job.json
+      if [[ "${POD_TAG}" == "${PREVIOUS_POD_TAG}" ]]; then
+        echo [warning] commit id is the same, will not build again!
+        exit 0
+      fi
       trap 'docker-compose --project-name "${POD_NAME}-${POD_TAG}" down -v --rmi all --remove-orphans' EXIT
       docker-compose --project-name "${POD_NAME}-${POD_TAG}" --no-ansi build --no-cache
       docker-compose --project-name "${POD_NAME}-${POD_TAG}" --no-ansi push
